# autopilot-api/src/modules/tenants/dto/update-tenants.dto.ts
import {
  IsString,
  IsOptional,
  IsUUID,
  IsDate,
  IsObject,
} from 'class-validator';

export class TenantsUpdateDto {
  @IsOptional()
  @IsString()
  name?: string;

  @IsOptional()
  @IsString()
  slug?: string;

  @IsOptional()
  @IsString()
  logoUrl?: string;

  @IsOptional()
  @IsUUID()
  ownerId?: string;

  @IsOptional()
  @IsObject()
  themeConfig?: Record<string, unknown>;

  @IsOptional()
  @IsDate()
  createdAt?: Date;

  @IsOptional()
  @IsDate()
  updatedAt?: Date;
}